+++ src/demo/re_surface/surfaces/surface_18.cljs
@@ -148,15 +148,12 @@
   {:header           {:key    :surface-18-header
                       :height 100
                       :fixed? true}
-   :header-dropdown {:key :surface-18-header-dropdown
-                     :active? false
-                     :height 250
+   :header-dropdown {:active? false
                      :full-width? true}
    :navbar           {:key    :surface-18-navbar
                       :height 100
                       :fixed? true}
-   :navbar-dropdown {:key     :surface-18-navbar-dropdown
-                     :active? false
+   :navbar-dropdown {:active? false
                      :top     -24}
    :body             {:key :surface-18-body}
    :footer           {:key    :default
@@ -169,12 +166,10 @@
                       :width   200
                       :active? false}
    :dimmer           {:key :surface-18-dimmer}
-   :modal            {:key     :surface-18-modal
-                      :active? false
+   :modal            {:active? false
                       :width   300
                       :height  500}
-   :modal-fullscreen {:key        :surface-18-modal-fs
-                      :active?    false
+   :modal-fullscreen {:active?    false
                       }})
 
 (def surfaces
@@ -188,25 +183,32 @@
    :surface-18-sra (-> surface-init
                        (assoc-in [:sidebar-right :active?] true))
 
-
    :surface-18-modal (-> surface-init
+                         (assoc-in [:modal :key] :surface-18-modal)
                          (assoc-in [:modal :active?] true))
 
    ;; fs --> fullscreen
    :surface-18-modal-fs (-> surface-init
+                            (assoc-in [:modal-fullscreen :key]
+                                      :surface-18-modal-fs)
                             (assoc-in [:modal-fullscreen :active?] true))
 
    :surface-18-modal-fs-with-modal-on-top
    (-> surface-init
+       (assoc-in [:modal-fullscreen :key]
+                 :surface-18-modal-fs)
        (assoc-in [:modal-fullscreen :active?] true)
        (assoc-in [:dimmer :key] :surface-18-dimmer-with-modal-on-top)
+       (assoc-in [:modal :key] :surface-18-modal)
        (assoc-in [:modal :active?] true))
 
    :surface-18-header-dropdown
    (-> surface-init
+       (assoc-in [:header-dropdown :key] :surface-18-header-dropdown)
        (assoc-in [:header-dropdown :active?] true))
 
    :surface-18-navbar-dropdown
    (-> surface-init
+       (assoc-in [:navbar-dropdown :key] :surface-18-navbar-dropdown)
        (assoc-in [:navbar-dropdown :active?] true))
    })