+++ src/demo/re_surface/surfaces/surface_19.cljs
@@ -145,16 +145,12 @@
   {:header           {:key    :surface-19-header
                       :height 100
                       :fixed? true}
-   :header-dropdown  {:key         :surface-19-header-dropdown
-                      :active?     false
-                      :height      250
+   :header-dropdown  {:active?     false
                       :full-width? true}
    :navbar           {:key    :surface-19-navbar
                       :height 100
                       :fixed? true}
-   :navbar-dropdown  {:key         :surface-19-navbar-dropdown
-                      :active?     false
-                      :height      250
+   :navbar-dropdown  {:active?     false
                       :full-width? true}
    :body             {:key :surface-19-body}
    :footer           {:key    :default
@@ -167,13 +163,10 @@
                       :width   200
                       :active? false}
    :dimmer           {:key :surface-19-dimmer}
-   :modal            {:key     :surface-19-modal
-                      :active? false
+   :modal            {:active? false
                       :width   300
                       :height  500}
-   :modal-fullscreen {:key     :surface-19-modal-fs
-                      :active? false
-                      }})
+   :modal-fullscreen {:active? false}})
 
 (def surfaces
   {:surface-19 surface-init
@@ -186,25 +179,32 @@
    :surface-19-sra (-> surface-init
                        (assoc-in [:sidebar-right :active?] true))
 
-
    :surface-19-modal (-> surface-init
+                         (assoc-in [:modal :key] :surface-19-modal)
                          (assoc-in [:modal :active?] true))
 
    ;; fs --> fullscreen
    :surface-19-modal-fs (-> surface-init
+                            (assoc-in [:modal-fullscreen :key]
+                                      :surface-19-modal-fs)
                             (assoc-in [:modal-fullscreen :active?] true))
 
    :surface-19-modal-fs-with-modal-on-top
    (-> surface-init
+       (assoc-in [:modal-fullscreen :key]
+                 :surface-19-modal-fs)
        (assoc-in [:modal-fullscreen :active?] true)
        (assoc-in [:dimmer :key] :surface-19-dimmer-with-modal-on-top)
+       (assoc-in [:modal :key] :surface-19-modal)
        (assoc-in [:modal :active?] true))
 
    :surface-19-header-dropdown
    (-> surface-init
+       (assoc-in [:header-dropdown :key] :surface-19-header-dropdown)
        (assoc-in [:header-dropdown :active?] true))
 
    :surface-19-navbar-dropdown
    (-> surface-init
+       (assoc-in [:navbar-dropdown :key] :surface-19-navbar-dropdown)
        (assoc-in [:navbar-dropdown :active?] true))
    })