+++ src/main/re_surface/z_index.cljs
@@ -2,32 +2,45 @@
 
 
 (def indicies
-  {:dimmer          -1
-   :modal           -1
+  {:dimmer           -1
+   :modal            -1
    :modal-fullscreen -1
-   :header-dropdown -1
-   :navbar-dropdown -1
+   :header-dropdown  -1
+   :navbar-dropdown  -1
 
    :surface 0
    :main    0
    :body    0
 
-   :header 1
-   :navbar 1
-   :footer 1
-   :header-dropdown-active 1
-   :navbar-dropdown-active 1
+   :header                 1
+   :navbar                 1
+   :footer                 1
 
-   ;; TODO: notifications 2
+   ;; PLACEHOLDER 2
+   ;; PLACEHOLDER 3
 
-   :dimmer-active 3
+   :dimmer-with-dropdown 4
 
-   :sidebar-left  4
-   :sidebar-right 4
+   :header-dropdown-active 5
+   :navbar-dropdown-active 6
 
-   :modal-fullscreen-active 5
+   ;; PLACEHOLDER 7
+   ;; PLACEHOLDER 8
 
-   :dimmer-active-with-modal 6
+   :dimmer-with-sidebar 9
 
-   :modal-active 7
+   :sidebar-left  10
+   :sidebar-right 11
+
+   ;; PLACEHOLDER 12
+   ;; PLACEHOLDER 13
+
+   :modal-fullscreen-active 14
+
+   ;; PLACEHOLDER 15
+   ;; PLACEHOLDER 16
+
+   :dimmer-with-modal 17
+
+   :modal-active 18
    })